+++ cypress/support/commands.ts
@@ -204,6 +204,44 @@ Cypress.Commands.add(
   },
 );
 
+Cypress.Commands.add(
+  'checkCorrectNumberOfProfessionsAreShown',
+  (statuses: ('live' | 'archived' | 'draft')[]) => {
+    cy.readFile('./seeds/test/professions.json').then((professions) => {
+      const professionsToShow = professions.filter((profession) =>
+        profession.versions.some((version) =>
+          statuses.includes(version.status),
+        ),
+      );
+
+      cy.translate('professions.search.foundPlural', {
+        count: professionsToShow.length,
+      }).then((foundText) => {
+        cy.get('body').should('contain', foundText);
+      });
+    });
+  },
+);
+
+Cypress.Commands.add(
+  'checkCorrectNumberOfOrganisationsAreShown',
+  (statuses: ('live' | 'archived' | 'draft')[]) => {
+    cy.readFile('./seeds/test/organisations.json').then((organisations) => {
+      const organisationsToShow = organisations.filter((organisation) =>
+        organisation.versions.some((version) =>
+          statuses.includes(version.status),
+        ),
+      );
+
+      cy.translate('organisations.search.foundPlural', {
+        count: organisationsToShow.length,
+      }).then((foundText) => {
+        cy.get('body').should('contain', foundText);
+      });
+    });
+  },
+);
+
 Cypress.Commands.add('visitAndCheckAccessibility', (url: string) => {
   cy.visit(url);
   cy.checkAccessibility();