+++ src/professions/search/interfaces/index-template.interface.ts
@@ -13,4 +13,6 @@ export interface IndexTemplate {
   nationsCheckboxItems: CheckboxItems[];
 
   industriesCheckboxItems: CheckboxItems[];
+
+  hasSelectedFilters: boolean;
 }